requirements.txt
Package                  Version
------------------------ --------------------
absl-py                  1.4.0
altair                   4.2.2
asttokens                2.2.1
attrs                    23.1.0
backcall                 0.2.0
blinker                  1.6.2
cachetools               5.3.0
certifi                  2022.12.7
charset-normalizer       2.1.1
click                    8.1.3
colorama                 0.4.6
contourpy                1.0.7
cycler                   0.11.0
decorator                5.1.1
entrypoints              0.4
executing                1.2.0
filelock                 3.9.0
fonttools                4.39.4
gcloud                   0.17.0
gitdb                    4.0.10
GitPython                3.1.31
google-auth              2.18.0
google-auth-oauthlib     1.0.0
googleapis-common-protos 1.59.0
grpcio                   1.54.0
httplib2                 0.22.0
idna                     3.4
importlib-metadata       6.6.0
importlib-resources      5.12.0
ipython                  8.13.2
jedi                     0.18.2
Jinja2                   3.1.2
jsonschema               4.17.3
jws                      0.1.3
kiwisolver               1.4.4
Markdown                 3.4.3
markdown-it-py           2.2.0
MarkupSafe               2.1.2
matplotlib               3.7.1
matplotlib-inline        0.1.6
mdurl                    0.1.2
mpmath                   1.2.1
networkx                 3.0
numpy                    1.24.1
oauth2client             3.0.0
oauthlib                 3.2.2
opencv-python            4.7.0.72
opencv-python-headless   4.7.0.72
packaging                23.1
pandas                   2.0.1
parso                    0.8.3
pickleshare              0.7.5
Pillow                   9.3.0
pip                      21.2.4
prompt-toolkit           3.0.38
protobuf                 3.20.3
psutil                   5.9.5
pure-eval                0.2.2
pyarrow                  12.0.0
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pycryptodome             3.4.3
pydeck                   0.8.1b0
Pygments                 2.15.1
Pympler                  1.0.1
pyparsing                3.0.9
Pyrebase                 3.0.27
pyrsistent               0.19.3
python-dateutil          2.8.2
python-jwt               2.0.1
pytz                     2023.3
pytz-deprecation-shim    0.1.0.post0
PyYAML                   6.0
requests                 2.28.1
requests-oauthlib        1.3.1
requests-toolbelt        0.7.0
rich                     13.3.5
rsa                      4.9
scipy                    1.10.1
seaborn                  0.12.2
setuptools               58.1.0
six                      1.16.0
smmap                    5.0.0
st-btn-select            0.1.2
stack-data               0.6.2
streamlit                1.22.0
sympy                    1.11.1
tenacity                 8.2.2
tensorboard              2.13.0
tensorboard-data-server  0.7.0
thop                     0.1.1.post2209072238
toml                     0.10.2
toolz                    0.12.0
torch                    2.0.1+cu117
torchaudio               2.0.2+cu117
torchvision              0.15.2+cu117
tornado                  6.3.1
tqdm                     4.65.0
traitlets                5.9.0
typing_extensions        4.4.0
tzdata                   2023.3
tzlocal                  4.3
urllib3                  1.26.13
validators               0.20.0
watchdog                 3.0.0
wcwidth                  0.2.6
Werkzeug                 2.3.4
wget                     3.2
wheel                    0.40.0
zipp                     3.15.0
